Update code: 1. Fix random crash on Ubuntu 2. Add Ubuntu docker 3. Add progress to clinical counter in ValidationReporter

+++ src/Utils/TypeDef.h
@@ -20,7 +20,7 @@ class IStrategy;
 
 class Therapy;
 
-using Ul = unsigned int;
+using Ul = uint64_t;
 
 using DoubleVector = std::vector<double>;
 using DoubleVector2 = std::vector<DoubleVector>;
